remove /save-fhir-data and all references
