+++ containers/ecr-viewer/tests/integration/saveFhirDataService/core.test.ts
@@ -4,8 +4,7 @@
 
 import { getLastAuditLog } from "../helpers/core";
 import { buildCore, clearCore, dropExisting } from "../helpers/ddl";
-import { saveFhirMetadata } from "@/app/api/save-fhir-data/service";
-import { BundleMetadata } from "@/app/api/save-fhir-data/types";
+import { BundleMetadata, saveFhirMetadata } from "@/app/services/saveFhirDataService";
 import { BlobResponse } from "@/app/data/blobStorage/utils";
 import { getDb } from "@/app/data/metadataDb/database";
 import { Core } from "@/app/data/metadataDb/types/core";
