+++ containers/ecr-viewer/tests/integration/saveFhirDataService/extended.test.ts
@@ -4,8 +4,7 @@
 
 import { getLastAuditLog } from "../helpers/core";
 import { buildExtended, clearExtended, dropExisting } from "../helpers/ddl";
-import { saveFhirMetadata } from "@/app/api/save-fhir-data/service";
-import { BundleExtendedMetadata } from "@/app/api/save-fhir-data/types";
+import { BundleExtendedMetadata, saveFhirMetadata } from "@/app/services/saveFhirDataService";
 import { BlobResponse } from "@/app/data/blobStorage/utils";
 import { getDb } from "@/app/data/metadataDb/database";
 import { Extended } from "@/app/data/metadataDb/types/extended";
