+++ containers/orchestration/.env.sample
@@ -2,7 +2,6 @@
 FHIR_CONVERTER_URL=http://fhir-converter-proxy-service:8080
 INGESTION_URL=http://ingestion-service:8080
 MESSAGE_PARSER_URL=http://message-parser-service:8080
-ECR_VIEWER_URL=http://host.docker.internal:3000/ecr-viewer
 TRIGGER_CODE_REFERENCE_URL=http://trigger-code-reference-service:8080
 
 # Use the FHIR Converter Proxy Port here (8087). If not using a proxy, set to: 8082
@@ -10,9 +9,6 @@ fhir_converter_url=http://localhost:8087
 ingestion_url=http://localhost:8083
 message_parser_url=http://localhost:8085
 trigger_code_reference_url=http://localhost:8086
-ecr_viewer_url=http://localhost:3000/ecr-viewer
-
-DATABASE_URL=postgresql://postgres:pw@db:5432/ecr_viewer_db
 
 ORCHESTRATION_PORT_NUMBER="8080"
 FHIR_CONVERTER_PORT_NUMBER="8082"
@@ -20,5 +16,3 @@ INGESTION_PORT_NUMBER="8083"
 MESSAGE_PARSER_PORT_NUMBER="8085"
 TRIGGER_CODE_REFERENCE_PORT_NUMBER="8086"
 FHIR_CONVERTER_PROXY_PORT_NUMBER="8087"
-ECR_VIEWER_HEALTH_CHECK="3000/ecr-viewer/api/health-check"
-DB_PORT_NUMBER="5432"
