--- containers/orchestration/app/custom_configs/integrated.json
@@ -1,35 +1,0 @@
-{
-  "workflow": [
-    {
-      "service": "fhir_converter",
-      "endpoint": "/convert-to-fhir"
-    },
-    {
-      "service": "ingestion",
-      "endpoint": "/fhir/harmonization/standardization/standardize_names"
-    },
-    {
-      "service": "ingestion",
-      "endpoint": "/fhir/harmonization/standardization/standardize_dob",
-      "params": {
-        "dob_format": ""
-      }
-    },
-    {
-      "name": "fhir_bundle",
-      "service": "ingestion",
-      "endpoint": "/fhir/harmonization/standardization/standardize_phones"
-    },
-    {
-      "name": "stamped_ecr",
-      "service": "trigger_code_reference",
-      "endpoint": "/stamp-condition-extensions"
-    },
-    {
-      "name:": "save_bundle",
-      "service": "save_bundle",
-      "url": "${ECR_VIEWER_URL}",
-      "endpoint": "/api/save-fhir-data"
-    }
-  ]
-}
