--- containers/orchestration/app/custom_configs/non-integrated-extended.json
@@ -1,51 +1,0 @@
-{
-  "workflow": [
-    {
-      "service": "fhir_converter",
-      "endpoint": "/convert-to-fhir"
-    },
-    {
-      "service": "ingestion",
-      "endpoint": "/fhir/harmonization/standardization/standardize_names"
-    },
-    {
-      "service": "ingestion",
-      "endpoint": "/fhir/harmonization/standardization/standardize_dob",
-      "params": {
-        "dob_format": ""
-      }
-    },
-    {
-      "name": "fhir_bundle",
-      "service": "ingestion",
-      "endpoint": "/fhir/harmonization/standardization/standardize_phones"
-    },
-    {
-      "name": "stamped_ecr",
-      "service": "trigger_code_reference",
-      "endpoint": "/stamp-condition-extensions"
-    },
-    {
-      "name": "metadata_values",
-      "service": "message_parser",
-      "endpoint": "/parse_message",
-      "params": {
-        "message_format": "fhir",
-        "parsing_schema_name": "extended.json",
-        "credential_manager": "azure"
-      }
-    },
-    {
-      "name:": "save_bundle",
-      "service": "save_bundle",
-      "url": "${ECR_VIEWER_URL}",
-      "endpoint": "/api/save-fhir-data",
-      "previous_response_to_param_mapping": {
-        "metadata_values": "metadata",
-        "stamped_ecr": "fhirBundle"
-      }
-    }
-  ],
-  "outputs": ["metadata_values", "stamped_ecr"],
-  "default-response": false
-}
