--- containers/orchestration/app/custom_configs/save-bundle-to-ecr-viewer.json
@@ -1,23 +1,0 @@
-{
-  "workflow": [
-    {
-      "name": "metadata_values",
-      "service": "message_parser",
-      "endpoint": "/parse_message",
-      "params": {
-        "message_format": "fhir",
-        "parsing_schema_name": "core.json"
-      }
-    },
-    {
-      "name:": "save_bundle",
-      "service": "save_bundle",
-      "endpoint": "/api/save-fhir-data",
-      "previous_response_to_param_mapping": {
-        "metadata_values": "metadata"
-      }
-    }
-  ],
-  "outputs": ["metadata_values"],
-  "default-response": false
-}
