+++ containers/orchestration/app/default_configs/sample-orchestration-s3-config.json
@@ -24,14 +24,6 @@
       "endpoint": "/fhir/harmonization/standardization/standardize_phones"
     },
     {
-      "service": "save_bundle",
-      "url": "${ECR_VIEWER_URL}",
-      "endpoint": "/api/save-fhir-data",
-      "params": {
-        "saveSource": "s3"
-      }
-    },
-    {
       "service": "message_parser",
       "url": "${MESSAGE_PARSER_URL}",
       "endpoint": "/parse_message",
