+++ containers/orchestration/local-dev.env
@@ -4,15 +4,9 @@ INGESTION_URL=http://localhost:8083
 MESSAGE_PARSER_URL=http://localhost:8085
 TRIGGER_CODE_REFERENCE_URL=http://localhost:8086
 
-DATABASE_URL=postgresql://postgres:pw@localhost:5432/ecr_viewer_db
-
-AWS_REGION=us-east-1
-ECR_BUCKET_NAME=ecr-viewer-files
-
 ORCHESTRATION_PORT_NUMBER="8080"
 FHIR_CONVERTER_PORT_NUMBER="8082"
 INGESTION_PORT_NUMBER="8083"
 MESSAGE_PARSER_PORT_NUMBER="8085"
 TRIGGER_CODE_REFERENCE_PORT_NUMBER="8086"
 FHIR_CONVERTER_PROXY_PORT_NUMBER="8087"
-DB_PORT_NUMBER="5432"
