+++ containers/orchestration/pytest.ini
@@ -3,7 +3,6 @@ env =
     FHIR_CONVERTER_URL=http://localhost:8082
     INGESTION_URL=http://localhost:8083
     MESSAGE_PARSER_URL=http://localhost:8085
-    DATABASE_URL=postgresql://postgres:pw@localhost:5432/ecr_viewer_db
     SMARTY_AUTH_ID=placeholder
     SMARTY_AUTH_TOKEN=placeholder
