+++ containers/orchestration/tests/test_handlers.py
@@ -33,7 +33,7 @@ def test_build_fhir_converter_request():
     orchestration_request = {
         "message_type": "ecr",
         "data_type": "ecr",
-        "config_file_name": "integrated.json",
+        "config_file_name": "bundle-only.json",
         "message": message,
     }
     result = build_fhir_converter_request(message, orchestration_request)
